+++ doc/python/pages/changelog.rst
@@ -35,6 +35,10 @@ Changelog
 `Changes since 2020.06`_
 ========================
 
+-   Renamed :py:`Matrix3.from()` / :py:`Matrix4.from()` to :ref:`Matrix3.from_()`
+    / :ref:`Matrix4.from_()` because :py:`from` is a Python keyword and it
+    would be silly to have to write :py:`getattr(Matrix4, 'from')` just to use
+    these APIs
 -   Exposed :ref:`gl.Renderer.set_blend_function()`,
     :ref:`gl.Renderer.set_blend_equation()` and related enums (see :gh:`mosra/magnum-bindings#9`)
 -   Exposed :ref:`gl.Renderer.Feature.CLIP_DISTANCEn <gl.Renderer.Feature.CLIP_DISTANCE0>`